# docs/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "CodeSage API",
    "description": "Local-first, stateless AI pull request review assistant.\n\n## Authentication\nNo user authentication. Configure `GitHub:PersonalAccessToken` and `OpenAI:ApiKey`\nvia user-secrets or environment variables.\n\n## Errors\nAll failures return RFC 7807 `application/problem+json` with an `errorCode` extension.\n\n## Versioning\nCurrent surface is `/api/v1`. Breaking changes will ship as `/api/v2`.",
    "contact": {
      "name": "CodeSage Contributors",
      "url": "https://github.com/codesage"
    },
    "license": {
      "name": "MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/repositories/{owner}/{name}/pull-requests": {
      "get": {
        "tags": [
          "Pull Requests"
        ],
        "summary": "List pull requests",
        "description": "Lists pull requests for a repository (all states, recently updated first).",
        "operationId": "ListPullRequestsV1",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PullRequestSummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error — RFC 7807 ProblemDetails.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{owner}/{name}/pull-requests/{number}": {
      "get": {
        "tags": [
          "Pull Requests"
        ],
        "summary": "Get pull request",
        "description": "Returns pull request details including files, commits, and review/issue comments.",
        "operationId": "GetPullRequestV1",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullRequestDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error — RFC 7807 ProblemDetails.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{owner}/{name}/pull-requests/{number}/analysis": {
      "get": {
        "tags": [
          "Pull Requests"
        ],
        "summary": "Analyze pull request",
        "description": "Builds a deterministic ReviewContext for the pull request. Does not call OpenAI. Pass the result to POST /api/v1/reviews for an AI review.",
        "operationId": "AnalyzePullRequestV1",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewContext"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error — RFC 7807 ProblemDetails.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "List repositories",
        "description": "Lists repositories visible to the configured GitHub Personal Access Token (owned, collaborator, and organization memberships the token can see).",
        "operationId": "ListRepositoriesV1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositorySummaryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — RFC 7807 ProblemDetails with an errors object.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error — RFC 7807 ProblemDetails.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{owner}/{name}": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get repository",
        "description": "Returns details for a single repository identified by owner and name.",
        "operationId": "GetRepositoryV1",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error — RFC 7807 ProblemDetails.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChangedFileDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "additions": {
            "type": "integer",
            "format": "int32"
          },
          "deletions": {
            "type": "integer",
            "format": "int32"
          },
          "changes": {
            "type": "integer",
            "format": "int32"
          },
          "patch": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "A file changed in a pull request."
      },
...
      "CommitSummaryDto": {
        "type": "object",
        "properties": {
          "sha": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "authorLogin": {
            "type": "string",
            "nullable": true
          },
          "committedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Commit summary associated with a pull request."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "PullRequestCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "authorLogin": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "kind": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "line": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Issue or review comment on a pull request."
      },
      "PullRequestDetailsDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string"
          },
          "draft": {
            "type": "boolean"
          },
          "authorLogin": {
            "type": "string"
          },
          "authorAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "htmlUrl": {
            "type": "string"
          },
          "baseRef": {
            "type": "string"
          },
          "headRef": {
            "type": "string"
          },
          "changedFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChangedFileDto"
            }
          },
          "commits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitSummaryDto"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PullRequestCommentDto"
            }
          }
        },
        "additionalProperties": false,
        "description": "Full pull request details including files, commits, and comments."
      },
      "PullRequestSummaryDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "draft": {
            "type": "boolean"
          },
          "authorLogin": {
            "type": "string"
          },
          "authorAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "htmlUrl": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "Compact pull request listing item."
      },
      "RepositoryDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "ownerLogin": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "private": {
            "type": "boolean"
          },
          "htmlUrl": {
            "type": "string"
          },
          "defaultBranch": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "openIssuesCount": {
            "type": "integer",
            "format": "int32"
          },
          "forksCount": {
            "type": "integer",
            "format": "int32"
          },
          "stargazersCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Detailed repository view."
      },
      "RepositorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "ownerLogin": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "private": {
            "type": "boolean"
          },
          "htmlUrl": {
            "type": "string"
          },
          "defaultBranch": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Compact repository listing item."
      },
      "ReviewAuthorInfo": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReviewContext": {
        "type": "object",
        "properties": {
          "repository": {
            "$ref": "#/components/schemas/ReviewRepositoryInfo"
          },
          "pullRequest": {
            "$ref": "#/components/schemas/ReviewPullRequestInfo"
          },
          "author": {
            "$ref": "#/components/schemas/ReviewAuthorInfo"
          },
          "commits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewCommitInfo"
            }
          },
          "changedFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewFileChange"
            }
          },
          "statistics": {
            "$ref": "#/components/schemas/ReviewStatistics"
          },
          "languageBreakdown": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "summary": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "The only object future AI providers should consume.\nContains no GitHub / Octokit / vendor types — only CodeSage vocabulary.",
        "example": "{\"repository\":{\"name\":\"Healthcare.API\",\"fullName\":\"acme/Healthcare.API\",\"defaultBranch\":\"main\"},\"pullRequest\":{\"number\":42,\"title\":\"Harden auth\",\"description\":\"...\",\"state\":\"open\",\"isDraft\":false,\"baseRef\":\"main\",\"headRef\":\"feature/auth\",\"createdAt\":\"2026-07-01T12:00:00Z\",\"updatedAt\":\"2026-07-02T09:00:00Z\"},\"author\":{\"login\":\"dev\",\"name\":null,\"avatarUrl\":null},\"commits\":[],\"changedFiles\":[],\"statistics\":{\"fileCount\":1,\"commitCount\":1,\"additions\":10,\"deletions\":2,\"totalChangedLines\":12,\"languagesUsed\":[\"C#\"],\"largestModifiedFile\":\"Auth.cs\",\"largestModifiedFileChanges\":12,\"testFilesChanged\":0,\"sqlFilesChanged\":0,\"configurationFilesChanged\":0,\"controllerFilesChanged\":0,\"serviceFilesChanged\":0,\"sqlModified\":false},\"languageBreakdown\":{\"C#\":1},\"summary\":\"Deterministic analysis summary\"}"
      },
...
      "ReviewRepositoryInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "defaultBranch": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}